+++ configs/classification/cifar100/mixups/basic/rx50_mixups_CE_none.py
@@ -38,7 +38,8 @@ model = dict(
 # optimizer
 optimizer = dict(type='SGD',
                  lr=0.1, momentum=0.9, weight_decay=0.0001)  # adjust wd={1e-4, 5e-4} for mixup methods
-# apex
+
+# fp16
 use_fp16 = False
 optimizer_config = dict(update_interval=1, grad_clip=None)