+++ configs/classification/cifar100/optimizers/basic/r50_CE_none_bs100.py
@@ -1,0 +1,33 @@
+_base_ = [
+    '../../../_base_/datasets/cifar100/sz32_bs100.py',
+    '../../../_base_/default_runtime.py',
+]
+
+# model settings
+model = dict(
+    type='Classification',
+    pretrained=None,
+    backbone=dict(
+        type='ResNet_CIFAR',  # CIFAR
+        depth=50,
+        out_indices=(3,),  # no conv-1, x-1: stage-x
+        style='pytorch'),
+    head=dict(
+        type='ClsHead',  # normal CE loss (NOT SUPPORT PuzzleMix, use soft/sigm CE instead)
+        loss=dict(type='CrossEntropyLoss', loss_weight=1.0),
+        with_avg_pool=True, multi_label=False, in_channels=2048, num_classes=100)
+)
+
+# data
+data = dict(imgs_per_gpu=100, workers_per_gpu=4)
+
+# fp16
+use_fp16 = False
+fp16 = dict(type='mmcv', loss_scale='dynamic')
+optimizer_config = dict(grad_clip=None, update_interval=1)
+
+# learning policy
+lr_config = dict(policy='CosineAnnealing', min_lr=0.)
+
+# runtime settings
+runner = dict(type='EpochBasedRunner', max_epochs=200)