+++ configs/classification/cifar100/optimizers/basic/r50_CE_none_bs100_adam_ep200.py
@@ -1,0 +1,4 @@
+_base_ = 'r50_CE_none_bs100.py'
+
+# optimizer
+optimizer = dict(type='Adam', lr=1e-3)