+++ configs/classification/cifar100/optimizers/basic/r50_CE_none_bs100_adamw_ep200.py
@@ -1,0 +1,12 @@
+_base_ = 'r50_CE_none_bs100.py'
+
+# optimizer
+optimizer = dict(
+    type='AdamW',
+    lr=1e-3,
+    weight_decay=0.05, eps=1e-8, betas=(0.9, 0.999),
+    paramwise_options={
+        '(bn|ln|gn)(\d+)?.(weight|bias)': dict(weight_decay=0.),
+        'norm': dict(weight_decay=0.),
+        'bias': dict(weight_decay=0.),
+    })