+++ openmixup/core/optimizers/__init__.py
@@ -1,10 +1,16 @@
+from .adabelief import AdaBelief
+from .adafactor import Adafactor
+from .adahessian import Adahessian
+from .adamp import AdamP
 from .adan import Adan
 from .builder import build_optimizer
 from .constructor import DefaultOptimizerConstructor, TransformerFinetuneConstructor
 from .lamb import LAMB
 from .lars import LARS
+from .lion import Lion
+from .nadam import Nadam
 
 __all__ = [
-    'Adan', 'LARS', 'LAMB', 'build_optimizer',
-    'DefaultOptimizerConstructor', 'TransformerFinetuneConstructor'
+    'AdaBelief', 'Adafactor', 'Adahessian', 'AdamP', 'Adan', 'LARS', 'LAMB', 'Lion', 'Nadam',
+    'build_optimizer', 'DefaultOptimizerConstructor', 'TransformerFinetuneConstructor'
 ]